server/requirements.txt
fastapi
uvicorn
websockets
scikit-learn
joblib
numpy
google-generativeai
fastf1
pandas
